requirements.txt
absl-py==0.12.0
aio-pika==6.8.0
aiofiles==0.6.0
aiohttp==3.6.3
aiormq==3.3.1
altair==4.1.0
APScheduler==3.7.0
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
base58==2.1.0
beautifulsoup4==4.9.3
bidict==0.21.2
bleach==3.3.0
blinker==1.4
blis==0.7.4
boto3==1.17.58
botocore==1.20.58
cachetools==4.2.1
catalogue==2.0.4
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
colorclass==2.2.0
coloredlogs==14.3
colorhash==1.0.3
configparser==5.0.2
cryptography==3.4.7
cycler==0.10.0
cymem==2.0.5
Cython==0.29.21
datasets==1.6.0
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.3
distro==1.5.0
dm-tree==0.1.5
dnspython==1.16.0
docker-pycreds==0.4.0
docopt==0.6.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
entrypoints==0.3
et-xmlfile==1.1.0
fbmessenger==6.0.0
filelock==3.0.12
fsspec==2021.4.0
future==0.18.2
gast==0.3.3
gensim==4.0.1
gitdb==4.0.7
GitPython==3.1.14
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.37.0
h11==0.9.0
h5py==2.10.0
httpcore==0.11.1
httplib2==0.19.1
httptools==0.2.0
httpx==0.15.4
huggingface-hub==0.0.8
humanfriendly==9.1
idna==2.10
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==0.15.1
jsonpickle==2.0.0
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kafka-python==2.0.2
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
mattermostwrapper==2.2
mistune==0.8.4
multidict==4.7.6
multiprocess==0.70.11.1
murmurhash==1.0.5
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
nltk==3.6.2
notebook==6.3.0
numpy==1.18.5
oauth2client==4.1.3
oauthlib==3.1.0
openpyxl==3.0.7
opt-einsum==3.3.0
packaging==20.9
pamqp==2.3.0
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pathtools==0.1.2
pathy==0.5.2
pickleshare==0.7.5
Pillow==8.2.0
preshed==3.0.5
prometheus-client==0.10.1
promise==2.3
prompt-toolkit==2.0.10
protobuf==3.15.8
psutil==5.8.0
psycopg2-binary==2.8.6
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.7.4
pydeck==0.6.2
pydot==1.4.2
Pygments==2.8.1
PyJWT==2.0.1
pykwalify==1.8.0
pylev==1.4.0
pymongo==3.10.1
pyparsing==2.4.7
pyreadline==2.1
pyrsistent==0.17.3
pyTelegramBotAPI==3.7.7
python-crfsuite==0.9.7
python-dateutil==2.8.1
python-engineio==4.1.0
python-Levenshtein==0.12.2
python-socketio==5.2.1
pytz==2020.5
pywin32==300
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.0.3
questionary==1.9.0
rasa==2.5.0
rasa-sdk==2.5.0
redis==3.5.3
regex==2020.9.27
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rocketchat-API==1.15.0
rsa==4.7.2
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
ruyaml==0.20.0
s3transfer==0.4.2
sacremoses==0.0.45
sanic==20.9.0
Sanic-Cors==0.10.0.post3
sanic-jwt==1.5.0
Sanic-Plugins-Framework==0.9.5
scikit-learn==0.24.1
scipy==1.6.2
selenium==3.141.0
Send2Trash==1.5.0
sentencepiece==0.1.95
sentry-sdk==0.19.5
seqeval==1.2.2
shortuuid==1.0.1
simpletransformers==0.61.4
six==1.15.0
sklearn-crfsuite==0.3.6
slackclient==2.9.3
smart-open==3.0.0
smmap==4.0.0
sniffio==1.2.0
soupsieve==2.2.1
spacy==3.0.6
spacy-legacy==3.0.5
SQLAlchemy==1.3.24
srsly==2.4.1
streamlit==0.80.0
subprocess32==3.5.4
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorboardX==2.2
tensorflow==2.3.2
tensorflow-addons==0.12.0
tensorflow-estimator==2.3.0
tensorflow-hub==0.10.0
tensorflow-probability==0.11.1
termcolor==1.1.0
terminado==0.9.4
terminaltables==3.1.0
testpath==0.4.4
thinc==8.0.3
threadpoolctl==2.1.0
tokenizers==0.10.2
toml==0.10.2
toolz==0.11.1
torch==1.8.1+cu111
torchaudio==0.8.1
torchvision==0.9.1+cu111
tornado==6.1
tqdm==4.49.0
traitlets==5.0.5
transformers==4.5.1
twilio==6.50.1
typeguard==2.12.0
typer==0.3.2
typing-extensions==3.7.4.3
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.4
validators==0.18.2
wandb==0.10.27
wasabi==0.8.2
watchdog==2.0.3
wcwidth==0.2.5
webencodings==0.5.1
webexteamssdk==1.6
websockets==8.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==2.0.1
xxhash==2.0.2
yarl==1.5.1
